requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cymem==2.0.6
daphne==3.0.2
Django==4.0.5
django-cors-headers==3.13.0
django-nextjs==2.1.4
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
frozenlist==1.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
multidict==6.0.2
murmurhash==1.0.7
numpy==1.22.4
packaging==21.3
pathy==0.6.1
preshed==3.0.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
thinc==8.0.17
tqdm==4.64.0
Twisted==22.4.0
txaio==22.2.1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
websockets==10.3
yarl==1.7.2
zope.interface==5.4.0
